Relax the cursor position requirement in getting doc/types
I found myself constantly have to move cursor backwards in order to check
the type or doc of an identifier which I've just typed in (or tab completed),
as my cursor will always be after that identifier in these cases. And after my
inquiry is done, I'll again have to move the cursor back to where it was.

It feels ergonomic if we can additionally allow the cursor to be right after
an identifier when a user requests its doc.

The patch relaxes a little more than the above use cases, in that if there are
other things (other than spaces) e.g. brackets between the cursor and the
last identifier, it will work the same. I feel this is a no harm addition and
brings a bit more convenience probably.

Signed-off-by: Zheng Li <dev@zheng.li>